Fix User-Agent
Ficed the user-agent when making requests to the server because with the old one NIU's server would believe you were using the Mainland China app and would push notification in Chinese.
This forces the UA to english and to an Overseas user, just like the app does for an italian account.